requirements.txt
alabaster==0.7.16
Babel==2.15.0
bandit==1.7.5
black==23.7.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
docker==6.1.3
docutils==0.20.1
exceptiongroup==1.2.1
flake8==6.1.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
imagesize==1.4.1
iniconfig==2.0.0
Jinja2==3.1.2
joblib==1.4.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
mypy==1.4.1
mypy-extensions==1.0.0
numpy==1.25.2
packaging==24.1
pandas==2.0.3
pathspec==0.12.1
pbr==6.0.0
platformdirs==4.2.2
pluggy==1.5.0
psutil==5.9.5
pycodestyle==2.11.1
pyflakes==3.1.0
Pygments==2.18.0
pytest==7.4.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
scikit-learn==1.3.0
scipy==1.14.0
six==1.16.0
smmap==5.0.1
snowballstemmer==2.2.0
Sphinx==7.1.2
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
stevedore==5.2.0
threadpoolctl==3.5.0
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
websocket-client==1.8.0
yara-python==4.3.1
